fix being able to favorite routes twice
